requirements.txt
requests
beautifulsoup4
faiss-cpu
openai
tqdm